tus-clients/python/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==22.1.0
azure-core==1.26.3
azure-cosmos==4.3.0
azure-storage-blob==12.9.0
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.1.1
Click==7.0
click-conf==0.0.1
click-stream==0.0.10
cryptography==36.0.0
cx-Freeze==6.14.2
elastic-transport==8.1.2
elasticsearch==8.3.3
elasticsearch-loader==0.6.0
frozenlist==1.3.3
future==0.18.2
idna==3.4
isodate==0.6.0
jmespath==1.0.1
kafka-python==2.0.2
macholib==1.16.2
msrest==0.6.21
multidict==6.0.3
Nuitka==1.4.2
oauthlib==3.1.1
pyaml==21.10.1
pycparser==2.21
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.0
six==1.16.0
tinydb==4.7.0
tus.py==1.3.4
tuspy==1.0.0
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
